Uploading Final Files of Project
The Last Modification of the final files of the XML Parser Project.

+++ json.cpp
@@ -16,29 +16,25 @@ using namespace std;
 /*
  * Description :
  * Checks if a data is numeric.
- * Time Complexity = O(n)
- * Space Complexity = O(1)
  */
 bool isNumeric(string data)  {
 
-	for(int i = 0 ; i < data.size() ; i++) {
-		if(data[i] >= '0' && data[i] <= '9')
-			continue;
-		return false;
-	}
-	return true;
+    for(int i = 0 ; i < (int)data.size() ; i++) {
+        if(data[i] >= '0' && data[i] <= '9')
+            continue;
+        return false;
+    }
+    return true;
 }
 
 
 /*
  * Description :
  * Checks if a data is boolean (true or false).
- * * Time Complexity = O(1)
- * Space Complexity = O(1)
  */
 bool isBoolean(string data)	{
 
-	return data == "true" || data == "false";
+    return data == "true" || data == "false";
 
 }
 
@@ -46,12 +42,10 @@ bool isBoolean(string data)	{
 /*
  * Description :
  * Checks if a data is equal to "null".
- * * Time Complexity = O(1)
- * Space Complexity = O(1)
  */
 bool isNull(string data){
 
-	return data == "null";
+    return data == "null";
 
 }
